structures refactors , split /note , /proof , and removal of /wasm
