+++ packages/wasm-utils/src/lib.rs
@@ -1,7 +1,10 @@
 #![allow(dead_code)]
 
+#[cfg(feature = "wee_alloc")]
+#[global_allocator]
+static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
+
 pub mod note;
 pub mod proof;
 pub mod types;
 mod utils;
-pub mod wasm;
